refactor(zubarev.arsenii/F0): encapsulate file system operations with error handling and update utility functions
refactor(vector-citer.hpp): update iterator to conform to C++20 standards and improve type safety

refactor(my_vector/vector-iter.hpp): update iterator to conform to C++20 standards and add missing operators

+++ zubarev.arsenii/F0/haffman.cpp
@@ -148,7 +148,7 @@ namespace zubarev
 
       uint8_t byte = *it;
 
-      for (size_t i = 7; i >= 0; --i) {
+      for (int i = 7; i >= 0; --i) {
         if (few_bits >= total_bits) {
           break;
         }
